refactor(writers): hoist grid width helpers into the shared engine
Move the pure column-width computation (placement, fractional and
auto width distribution, colspan floor, line measurement) from the
plain writer into the shared grid module so the rst writer can reuse
the same formulas.
